[chore] #16 jwt 관련 파일 수정
