+++ src/main/java/org/festimate/team/auth/infra/KakaoProperties.java
@@ -1,0 +1,15 @@
+package org.festimate.team.auth.infra;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Getter
+@Setter
+@Configuration
+@ConfigurationProperties(prefix = "kakao")
+public class KakaoProperties {
+    private String clientId;
+    private String redirectUri;
+}
